May 21/requirements.txt
flask[async]==3.0.2
pyautogen==0.2.14
pandas==2.2.1
matplotlib==3.8.3
python-dotenv==1.0.1
google-generativeai==0.3.2
werkzeug==3.0.1
asgiref==3.7.2